W_SCREEN::setFocus fixes
- Ensure focusLost() isn't called when setFocus() is called on the widget that already has focus
- Set the new focus widget before calling focusLost() so that any focusLost() handler can obtain the new widget that has focus

+++ lib/widget/widget.cpp
@@ -1473,12 +1473,20 @@ void W_SCREEN::setFocus(const std::shared_ptr<WIDGET> &widget)
 {
 	if (auto locked = psFocus.lock())
 	{
+		if (locked == widget)
+		{
+			return; // do nothing - no change
+		}
+		psFocus = widget;
 		if (bWidgetsInitialized) // do not call focusLost if widgets have already shutdown / are not initialized
 		{
 			locked->focusLost();
 		}
 	}
-	psFocus = widget;
+	else
+	{
+		psFocus = widget;
+	}
 }
 
 void WidgSetAudio(WIDGET_AUDIOCALLBACK Callback, SWORD HilightID, SWORD ClickedID, SWORD ErrorID)
